Fix keyboard shortcut display layout
The keyboard_shortcuts_keyword TextView in keyboard_shortcut_app_item
layout was wrap_content making it take all the space in case the text is
too long. The TextView is now rendered indide a KeyboardShortcutsLayout
which extends the RelativeLayout and sets the maxWidth of
keyboard_shortcuts_keyword TextView to 70% when onMeasure is called with
EXACT MeasureSpec.

Bug: 32980732
Fixes: 32980732
Test: Manual
Change-Id: I1250802af0a4ad011e842318384ca6fe5686989c

Former-commit-id: e2c893fc2d26fd2a3c1accb8fc78dddc8022e1bf