Move ProcessCpuTracker.init into CpuTracker thread
ProcessCpuTracker.init is slow to start, moving it out of main thread
will accelerate the boot up. ProcessCpuTracker.init is protected with
synchronized keyword, so once the CpuTracker thread is up and running,
we should not see any racing condition caused by this move.

This changes saves ~140ms starting time on marlin

Test: run test frameworks/base/tests/ActivityTests
Bug: 33681304
Change-Id: I260a57ca721999be3a84c0e43837d684e47e3941

Former-commit-id: a5c336c32dae4586ff88d047a6def87b23d83d26