Don't move-to-front overlays for unresumed tasks
If we do move those tasks to the front, it will annoy the user of the
device because clearly that's not the activity they were using at the
time (otherwise it would have been resumed already)

This is a more generous check than the existing "if !top.visible" which
only prevented the new overlay activity from bringing the task to the
front if it was unresumed AND not visible on the screen.

Now it just has to be unresumed.

Bug: 31001762
Change-Id: I4576aaafa141ef5297f9c18935b747693a80f0fd

Former-commit-id: bee6422d7cbffde474aaac7d94bfffcaa0c9c126