Amend logic for sending power hint on activity start.
We previously used task boundaries to determine when to send the hint.
This does not properly capture the target use case. Additionally, the
implementation does not identify this condition properly.

The new check is based on comparing the process of the existing resumed
activity in the stack. We will send the hint if the process is changing
or there is no resumed activity.

Bug:32640289
Test: Manual - ran systrace capture script and verified regression
       resolved (observed consistently over multiple attempts).
Change-Id: Ie3cdd17773f3466d9b3c7abb6ec5b4093ddc3631

Former-commit-id: 834abec877ab33ce189d1c7ff53edcc64fe468f0