Show work unlock prompt on focus as well as touch
With change I4576aaafa141ef5297f9c18935b747693a80f0fd we can be
reasonably sure this isn't going to instantly bring the prompt into
fullscreen, since launching this activity as an overlay in an
unfocused task won't bring it to the front.

But if it's launched in a focused task, it will get focused and
brought to the front, starting the unlock process. This saves us a
click, and user frustration, and is generally a good thing.

Bug: 31001762
Test: //tests/PoApi/src/com/google/android/afwtest/poapi/WorkChallengeTest (harness)
Change-Id: Ic5a33589e2b63d04e47c5f166da0077188cb55ea

Former-commit-id: ad0118a358250e2c53a3fdd057cc30910fc47c31