Workaround for leaked dim layer.
Tasks are some how leaking their dim layer. The root cause of Task
not properly cleaning-up its dim layer in Task.removeImmediately()
hasn't being identified, but the CL will prevent user devices from
getting into a bad state until we figure-out the cause for the leak.

Note that when this occurs the task has been completely removed from
the system and the only reference left to it is for the dim layer...

Bug: 34395537
Test: This is a workaround until the real problem is identified.
Change-Id: Id8d45edb3b6858be82e4f057ae3da480da4f4b59

Former-commit-id: f0a60a9c60d9973e10a859dd8c31b8f18adb5a0e