wifi: WifiConfiguration: add isHomeProviderNetwork field
This is a boolean field indicating if the given WiFi network is provided
by a Passpont home provider. This is useful for the Passpoint provisioning
app to determine if the current WiFi connection is allowed to perform
Passpoint subscription or policy updates when it needs to.

This field will be set by PasspointNetworkEvaluator when generating
a WifiConfiguration for a Passpoint network.

Bug: 34467212
Test: make update-api; make -j32
Change-Id: I47310cda078cbf1f2d1b0a9bf79f487230fb668f

Former-commit-id: 1379fd0b1ef952547934846f4b6c40780f885e47