Fix for endless loop in RelativeLayout.
  getRelatedView(int[] rules, int relation)

is stuck in the loop, when a view is GONE, and references itself.
There is already a check in:

  getSortedViews(View[] sorted, int... rules) {

to make sure there are no loops of ViewA -> ViewB -> ViewA

BUT, there was a change made in findRoots(int[] rulesFilter)

to "Remove exception throw when a view has a self dependency
inside a RelativeLayout." so this means we must allow this case.

Other change git commit sha: da3003e1d71d66a1c936489025f8db314a2a4588

Bug: http://b/android.com/231353
Change-Id: Icc26b86ebbb19e482c3afe6a39db996ec493586d

Former-commit-id: a354c4f1565cb69ab0858b586c4f2071a46b1105