Fixing crash in SystemUI fetching task description icons.
When we compose the TaskDescription for the task, we update it with
information from each activity from top down.  If an activity has a
saved icon, the filepath for that icon will be encoded with the user
id of the activity and saved into the task description.  However, if
that activity's user is different from the user id of the root
activity of the task, we will later try to fetch the icon with the
user id of the task which will cause an exception to be thrown.

Instead, because the task description describes the task and information
belonging to the task description is always fetched as the user of
that task, we can save the icon with the task's user encoded instead
of the activity's.

Bug: 34692261
Test: Launch primary user activity with a custom task description icon
      in a task that is rooted with a work profile activity, then restart
      before the primary user activity is finished.

Change-Id: I02200f20c4ed70f9b230418bae7b347fbef7fe87

Former-commit-id: c8408b8fa758e0d10271bbe29f0e397c662a3641