Disable RescueParty in debug/lab environments.
We need to strike a balance between rescuing devices of real users
in the field, while also not interfering with debugging or testing
use-cases.

This change uses "charging via USB" as a signal that the device is
under active development, or is in a lab environment.  We could try
checking for an adb connection, but UsbDebuggingManager isn't
started until much later in the boot, so BatteryManager is our best
chance of detecting this case.

Note that we continue to always recover "user" builds, and we never
recover "eng" builds.

Test: builds, boots, skips when plugged in over USB
Bug: 34722552
Change-Id: Ib095482a2ed84e8c5dfb3c71d5f9b88457c2b5e6

Former-commit-id: cdee83a6147bf1601424d5451c30f0dfb8d0e2a6