Do not return cryptic strings from ClipData.Item.coerceToText
Currently, when ClipData contains an URI of "content", "file" or
"android.resource" scheme, and the URI cannot be opened
as a stream for some reason, then ClipData.Item.coerceToText
(or coerceToHtmlText) return the URI's textual representation.

This representation is not useful for a user at all.

Returning an empty string instead.

Bug: 28929820
Test: Drag a file from Files app to any text input, no URI should be inserted
Change-Id: I21a23b673ef145a008427e9177403b5b1a41b898

Former-commit-id: e1f5e13c0fe8a436f53b402361e43728d31d8888