Fix: fragment destroyed improperly during pop.
Bug 34850219

When multiple back stack records were popped together,
to the bottom of the stack, the fragment at the bottom
would be destroyed. This was caused by the back stack
being bumped prior to executing the operations. The
back stack record index didn't properly reflect the
state it should be in during the operation and the
final fragment returned false to isInBackStack().

This CL moves the back stack bumping to immediately
prior to executing the operations.

Test: I887d565d245cd61434963040c20f887a10bf51dc
Change-Id: I3ad2be65cf0e66bf5615c0e912cc233528ee72a3

Former-commit-id: 8c4fa3615dbc09a838f1a341e19c0d0e543f6150