Revert "Fix window transformation related issue"
This reverts commit 2af1abd69054cf1a67482384e738f845caaca258 [formerly 867210c1de62e533880da7ac64b8eaa837e6ed6a].

This commit is being reverted because the logic it tries to fix does not
make sense and caused a regression causing all animation positions to be
offset incorrectly but applying the animation metrics before applying the
window offset (mainly for bottom screen in multiwindow). The original
case this tries to fix is for older apps like gallery2 which crashes
(ramdump) when trying reproduce the steps.

Fixes: 34769878
Test: ran original case that this tried to fix and played with recents
related to the new bug it created

Former-commit-id: 34a06d148ec0218a5e065b7b9c7cf338c3d20df6