Zygote: Reset niceness if lazy preloading is enabled.
By definition, the zygote in question is lower priority if lazy
preloading is enabled. We therefore reset the zygote process'
niceness to 0 as soon as it comes up, rather than having it stick
around as a higher priority process until we eventually ask it
to load the default set of resources.

Test: manual.

Change-Id: I3ccfc0e7c3fdebf5f798d90b53140a554bf64c20

Former-commit-id: b49996d8e2658ea2d62cd19d28edd70b2b22f605