Force ephemeral apps to have correct targetSdk/targetSandbox temporarily
There is a bit of a chicken and egg problem with Instant Apps and the
requirement to have targerSdk >= O and targetSandboxVersion >= 2 with
the sdk not being locked down yet. To work around this for now the
targetSdk and targetSandboxVersion will be made correct instead of
causing the install to fail. This work around must be removed before O
ships.

Test: adb install --ephemeral updates the targetSdk and targetSandbox
accordingly.

Change-Id: I283e71e459f0f69ef4053591b6647991746f54da

Former-commit-id: d143808365c06be7e64837d005d07e91ebfad745