Move FragmentState out of Fragments.java [framework edition]
Fragments.java contained 2 top level classes FragmentState and Fragments.
This CL moves FragmentState into its own java file. On the plus side we
will have one top level class per file as mandated by Google style, on the
downside per line history will be a bit harder to do for FragmentState class,
however it does not seem to have a lot of logic so it should be an OK loss.

Test: None
Change-Id: I85a1e5e91c87c1b123026fa79efb96f1d65b4b82

Former-commit-id: 6acbe075f92a2a3d0269e877427ebeff2e5982a9