Partially revert commit 1bc47129b559885125d6a5d6ae1db1292e3cb3de [formerly ff5659fbc3026e621bdd926e96096e2957cfd42e].
ff5659fbc was intended to be a clean revert of d302634e52bbf37d3ad [formerly 660cf30e2100d8ed401]
that fixed a DOA issue. However, the author incorrectly rebased that
revert and picked up pieces of a separate change as well. This reverts
those sections of the change so that they can be submitted with the
correct attribution and with proper review.

Bug: 34415265
Bug: 34914977
Test: make

Change-Id: I977c3ba10bb05f04ac8cfd95d17954c454c83340

Former-commit-id: e714927b199a5454c8e944dbb04cb422b58e91d7