Hide logspam due to redundant finishComposingText
Currently we see two warning messages
  "finishComposingText on inactive InputConnection"
  "finishComposingText on inactive InputConnection"
every time every time the View focus is switched from one EditText
to another EditText on the same window, which is really spammy.

This is actually not critical if IInputConnectionWrapper was already
finished, because with my previous CL [1] it is guaranteed that
InputConnection#finishComposingText() was already called followed by
InputConnection#closeConnection(), which means that the connection
is closed and should not accept any further requests. Thus ignoring
further #finishComposingText() only means that the system and/or IME
is calling #finishComposingText() unnecessarily, which is worth
showing spammy warnings in production builds.

To reduce logspam this CL hides warnings from the above case behind
DEBUG flag.

 [1]: If2a03bc84d318775fd4a197fa43acde086eda442
      7d3b72911bd5ab9a2ab02ef6501885ba6769e499 [formerly aaa38c9f1ae019f0fe8c3ba80630f26e582cc89c]

Test: Make sure `adb logcat -s IInputConnectionWrapper:*` does not
      show "finishComposingText on inactive InputConnection" warnings
      while switching focus across different EditText on the same
      window.
Bug: 35079353
Bug: 35301295
Change-Id: I17f3a4f500bc19ebf8bae771bf658a93627b3ba3

Former-commit-id: 77e51831a448b03cbeeea8ea286b1d584dca74d4