Pinned stack animation: Trigger window preservation when dropping shadows.
When resizing from smaller to large we go in two phases. First we resize to large + insets
so that we can scale up without pixelation. Then once we have reached our target we remove
the insets. As pointed out in the previous CL we don't want the window preservation
mechanism for the first size change, but it is perfect for removing the shadows. Otherwise
we have an issue where the client jumps, as when the shadows are removed we need to update
the position in synchronization with the client drawing.

Bug: 35396882
Test: Move skeleton app to fullscreen and verify no position jump at end.
Change-Id: I87eaf866f5a5ed058844670eeadffc34c061a378

Former-commit-id: c91d1c3e5623490834495b12547a38313e89bb5b