Skip the MTP notification when the system has the default app.
Before N, when MTP device is connected to Android, it opened the
activity selector for applications that could handle the MTP device.
For N, we updated the behavior to make it consistent with other storage
devices like USB mass storage. In N, the system showed the notification
first, then showed the activity selector after a user taps the
notification.

The N behavior causes a problem that the user always needs to tap the
MTP notification even after the user select a default application for
the MTP device.

The CLs tweaks the behavior for O so that the system can skip the MTP
notification if the system has the default app for the connected MTP
device.

The default app is one of them:

 * App selected by a user by clicking "Always use" in the activity
   selector
 * System app that is an only application which can handle the MTP
   device in the system
 * App that is an only application which can handle the MTP device in
   the system where config_disableUsbPermissionDialogs=true.

Bug: 32119378
Test: Connecting MTP device to Android, which automatically opens system
      app. Connecting PTP device to Android, which shows a notification.
      Tapping the notifiction opens the acitivity selector.
Change-Id: I51f1d55e8c7d1db87d91ddf966631d275a171546

Former-commit-id: a455dd9f0939f36438b39c0ae6b69e4c82e6ff02