Fix build failure when using javac toolchain
The legacy-performance-test-hostdex library was built using the
BUILD_HOST_DALVIK_JAVA_LIBRARY macro (which creates a non-static
library) but is used with LOCAL_STATIC_JAVA_LIBRARIES which
requires a static library. That works with Jack but fails with
the javac toolchain.

This change switched to build legacy-performance-test-hostdex
using the BUILD_HOST_DALVIK_STATIC_JAVA_LIBRARY macro.

Bug: 30188076
Test: make checkbuild
Change-Id: I8d42c669e4ded3417ecdcb7b0251bc6346debd5d

Former-commit-id: f1c59610bbe5ca7bf28083a6aa9c6ef551dfde52