Add a debug flag to disable time correction.
This allows manual QA folks to change the datetime to test events
which depend on a large gap between uses of an app.

Currently, if the system detects a drift of more than 2 seconds,
it will automatically correct the usage stats. This means that manual
time changes will cause the usage stats to update to match, making it
impossible to test manually.

Bug: 34400961
Test: Manual
Change-Id: Iffb92c929872d841d22f089ec71922bf120cc544

Former-commit-id: 9acc4f40c8e31140c2034dc48b75d23dadf7e0c6