Use the measured height in the ResolverDrawer
This was done to eliminate some jank while the
service target are animating in due to some
inconsistent padding calculation. Tracking the root
cause there has been elusive, so we're temporarily
using the full height given we have bigger rewrites
we want of this component (e.g. migrating to
RecyclerView).

In the meanwhile this has the unwanted side effect of
letting the user maximize all drawers while pulling up.

bug: 34253328
Test: Manually verified behavior and lack of jank.

Change-Id: I3c5f52ed8180ac2da9e5c9f891879980e49728c0

Former-commit-id: 9dd9090affbb3bfbe2919bd9a10a613a5a250c2f