Compute merged configuration changes
When we compute configuration changes to decide whether to relaunch
an activity or just report a coniguration change, we were calculating
task override changes and global changes separately. This results in
incorrect set of detected changes if we're moving between the
displays. E.g. when moving from secondary display to the primary
a lot of override config values change from empty to non-empty.

This CL switches to calculating configuration changes for merged
task configuration (global + overrides).

Bug: 34164473
Test: android.server.cts.ActivityManagerDisplayTests
Test: #testOnMovedToDisplayCallback
Change-Id: Icd6d10a49ab3ef80e00caf92cfd46e491b079d50

Former-commit-id: 0d8ff25fabae95d5ffc826adbb68b4036d41af41