Compute merged configuration changes
When we compute configuration changes to decide whether to relaunch
an activity or just report a coniguration change, we were calculating
task override changes and global changes separately. This results in
incorrect set of detected changes if we're moving between the
displays. E.g. when moving from secondary display to the primary
a lot of override config values change from empty to non-empty.

This CL switches to calculating configuration changes for merged
task configuration (global + overrides).

In this second version switched to not marking change to undefined
value as an actual change when comparing configuration, so it will
behave similarly to Configuration#updateFrom. If new configuration
with undefined value is applied to an old one with corresponding
value already set, it won't be overwritten to undefined.

Bug: 34164473
Bug: 35911519
Test: android.server.cts.ActivityManagerDisplayTests
Test: #testOnMovedToDisplayCallback
Change-Id: Ie1b90ccd2181aa5fdab4f57b9856e7de6959a6cb

Former-commit-id: b43be0afff6c7ee4daaeb19058289e0ab66bf807